refactor: 整理檔案結構
