+++ src/components/navigation/infoMenuItem/InfoMenuActivity.vue
