+++ src/components/navigation/infoMenuItem/InfoMenuDetail.vue
