+++ src/router/index.ts
@@ -7,7 +7,7 @@ const routes: RouteRecordRaw[] = [
         meta: {
             isRequiresAuth: false
         },
-        component: () => import('@/components/views/pages/LoginPage.vue')
+        component: () => import('@/components/layout/LoginLayout.vue')
     },
     {
         path: '/404-page',
@@ -15,7 +15,7 @@ const routes: RouteRecordRaw[] = [
         meta: {
             isRequiresAuth: false
         },
-        component: () => import('@/components/views/pages/404Page.vue')
+        component: () => import('@/components/views/404Page.vue')
     },
     {
         path: '/privacy-policy',
@@ -23,7 +23,7 @@ const routes: RouteRecordRaw[] = [
         meta: {
             isRequiresAuth: false
         },
-        component: () => import('@/components/views/pages/PrivacyPolicyPage.vue')
+        component: () => import('@/components/views/PrivacyPolicyPage.vue')
     },
     {
         path: '/pub-server',
@@ -31,7 +31,7 @@ const routes: RouteRecordRaw[] = [
         meta: {
             isRequiresAuth: false
         },
-        component: () => import('@/components/views/pages/PubServerPage.vue')
+        component: () => import('@/components/views/PubServerPage.vue')
     },
     {
         path: '/drive',
@@ -39,23 +39,23 @@ const routes: RouteRecordRaw[] = [
         meta: {
             isRequiresAuth: true
         },
-        component: () => import('@/components/views/pages/DriveHomePage.vue'),
+        component: () => import('@/components/layout/MainLayout.vue'),
         redirect: '/drive/my-drive',
         children: [
             {
                 path: 'my-drive',
                 name: 'my-drive',
-                component: () => import('@/components/views/HomePageView/MyDrive.vue')
+                component: () => import('@/components/views/MyDrivePage.vue')
             },
             {
                 path: 'folders/:folderId',
                 name: 'my-drive-folders',
-                component: () => import('@/components/views/HomePageView/MyDrive.vue')
+                component: () => import('@/components/views/MyDrivePage.vue')
             },
             {
                 path: 'search',
                 name: 'search-result-page',
-                component: () => import('@/components/views/HomePageView/SearchResult.vue')
+                component: () => import('@/components/views/SearchResultPage.vue')
             }
         ]
     },
